Think it's better to have the CRUD service separated into data type. Database creation class and SQL.
Added some SQL that should allow an update if something exists, and an insert if it doesn't. This should probably only be used on importing, but I think it'll work well to avoid a whole bunch of database commands when you're adding tweets (The thinking is that you have to add the status if it doesn't exist, and then add the user and place ifthey don't exist. Checking for existence and all that would require a lot of events if you didn't have one big SQL script.)

Changed ordering of commands so we're sure the database and the model exists before the services start fetching data. The example I'm using as a guide for this has it Service -> Database -> Model, but that order didn't work for me without bombing on injecting the model.